requirements.txt

flask==3.0.2
flask-login==0.6.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.2.2
nltk==3.8.1
scikit-learn==1.4.2
openai==1.30.1
